requirements.txt
google-cloud-documentai
google-cloud-dlp
streamlit
google-generativeai
